v1/requirements.txt
atproto>=0.0.40
python-dotenv>=1.0.0
feedparser>=6.0.10
requests>=2.28.0
